add blocks on demand tests
